Move hard-coded constraints
