+++ executor/src/witgen/processor.rs
@@ -10,7 +10,10 @@ use crate::witgen::{query_processor::QueryProcessor, util::try_to_simple_poly, C
 
 use super::{
     affine_expression::AffineExpression,
-    data_structures::{column_map::WitnessColumnMap, copy_constraints::CopyConstraints, finalizable_data::FinalizableData},
+    data_structures::{
+        column_map::WitnessColumnMap, copy_constraints::CopyConstraints,
+        finalizable_data::FinalizableData,
+    },
     identity_processor::IdentityProcessor,
     rows::{CellValue, Row, RowIndex, RowPair, RowUpdater, UnknownStrategy},
     Constraints, EvalError, EvalValue, FixedData, IncompleteCause, MutableState, QueryCallback,
@@ -34,6 +37,48 @@ impl<'a, T: FieldElement> OuterQuery<'a, T> {
 
     pub fn is_complete(&self) -> bool {
         self.left.iter().all(|l| l.is_constant())
+    }
+}
+
+pub fn make_copy_constraints<T: FieldElement>(
+    fixed_data: &FixedData<T>,
+    witness_cols: &HashSet<PolyID>,
+) -> CopyConstraints {
+    let is_example = witness_cols
+        .iter()
+        .all(|c| fixed_data.column_name(c).starts_with("main_pythagoras."));
+
+    if is_example {
+        log::info!("Using hard-coded copy constraints for Pythagoras example");
+        // Hard-code copy constraints every 4 rows:
+        // - a[0] = a[1]
+        // - a[0] = b[1]
+        // - b[0] = a[2]
+        // - b[0] = b[2]
+        // - c[1] = a[3]
+        // - c[2] = b[3]
+        // - c[0] = c[3]
+        let a = fixed_data.try_column_by_name("main_pythagoras.a").unwrap();
+        let b = fixed_data.try_column_by_name("main_pythagoras.b").unwrap();
+        let c = fixed_data.try_column_by_name("main_pythagoras.c").unwrap();
+        let d = fixed_data.degree;
+        let index =
+            |block_index, local_index| RowIndex::from_degree(block_index * 4 + local_index, d);
+
+        let mut constraints = Vec::new();
+
+        for i in 0..d / 4 {
+            constraints.push(((a, index(i, 0)), (a, index(i, 1))));
+            constraints.push(((a, index(i, 0)), (b, index(i, 1))));
+            constraints.push(((b, index(i, 0)), (a, index(i, 2))));
+            constraints.push(((b, index(i, 0)), (b, index(i, 2))));
+            constraints.push(((c, index(i, 1)), (a, index(i, 3))));
+            constraints.push(((c, index(i, 2)), (b, index(i, 3))));
+            constraints.push(((c, index(i, 0)), (c, index(i, 3))));
+        }
+        CopyConstraints::new(constraints)
+    } else {
+        CopyConstraints::default()
     }
 }
 
@@ -94,7 +139,7 @@ impl<'a, 'b, 'c, T: FieldElement, Q: QueryCallback<T>> Processor<'a, 'b, 'c, T, 
             outer_query: None,
             inputs: Vec::new(),
             previously_set_inputs: BTreeMap::new(),
-            copy_constraints: CopyConstraints::from_fixed_data(fixed_data, witness_cols),
+            copy_constraints: make_copy_constraints(fixed_data, witness_cols),
         }
     }
